server/requirements.txt
opencv-python
deepface
tf_keras
fer
python-dotenv
pymongo
flask
flask-socketio
eventlet